// web/src/client/styles/palette.ts
export const enum PALETTE {
  BACKGROUND_MAIN = '#f6f8f9',
  BACKGROUND_LIGHT_GRAY = '#f1f1f1',
  BACKGROUND_GRAY_MID = '#cfd6e4',
  BACKGROUND_HOVER = '#e1e4e7',
  BACKGROUND_HOVER_LIGHT = '#edeef0',
  BORDER_GRAY = '#e8ecee',
  BORDER_MAIN_GRAY = '#d1d5da',
  BORDER_HOVER = '#afa3e3',
  BOX_SHADOW_PRIMARY = '#e7e1f6',
  ERROR_MAIN = '#ff4a00',
  LINK = '#0389ff',
  PLACEHOLDER = '#b9b9b9',
  PRIMARY_MAIN_DARK = '#4824eb',
  PRIMARY_MAIN = '#6648ee',
  PRIMARY_MAIN_LIGHT = '#886df4',
  PRIMARY_MAIN_LIGHTER = '#a998F6',
  PRIMARY_MAIN_LIGHTEST = '#f1eefa',
  PRIMARY_GREEN_DARK = '#00943b',
  PRIMARY_GREEN = '#00ad45',
  PRIMARY_YELLOW = '#fedb41',
  TEXT_MAIN = '#24292e',
  TEXT_GRAY = '#7e8499'
}

export const BackgroundColors = [
  '#FF8282',
  '#FFAF82',
  '#FFD482',
  '#A9D6AE',
  '#AFD8E1',
  '#829DFF',
  '#846EF3',
  '#C590DE'
]