// web/src/client/styles/typography.ts
export const enum FONT_FAMILY {
  SANS_SERIF = 'Lambdas, -apple-system, BlinkMacSystemFont, Roboto, "Helvetica Neue", Helvetica, Arial, sans-serif',
  MONOSPACE = 'SF Mono, Menlo, Monaco, Consolas, "Courier New", monospace'
}

export const enum ICON_SIZE {
  MD16 = '16px',
  MD18 = '18px',
  MD22 = '22px',
  MD24 = '24px',
  MD36 = '36px',
  MD48 = '48px'
}